requirements.txt
beautifulsoup
bs4
slacker
pyexcel
python-dateutil
selenium
xmltodict
urllib3
tqdm
pymysql
requests
certifi==2019.6.16
Click==7.0
et-xmlfile==1.0.1
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
openpyxl==2.6.2
Werkzeug==0.15.6
